Move Harness to a module in a new ops.testing package.
Also, make it possible to pip install ops[testing] to get Scenario.
